test(pi): split profile resource coverage
