+++ tests/unit/pi-adapter.test.ts
@@ -445,41 +445,6 @@ describe('pi adapter', () => {
     });
   });
 
-  // THIS TEST VALIDATES A HARD REQUIREMENT (OFTR-006.3).
-  // YOU MUST NOT MODIFY THIS TEST UNLESS THE REQUIREMENT CHANGES.
-  it('adds profile-bundled Pi skills to the launch args', () => {
-    const root = createTemporaryPiAdapterTestRoot('outfitter-pi-profile-skills-');
-    const profileFolder = join(root, 'profiles', 'data_analyst');
-    const skillFolder = join(profileFolder, 'skills', 'demos');
-    const incompleteSkillFolder = join(profileFolder, 'skills', 'draft');
-    mkdirSync(skillFolder, { recursive: true });
-    mkdirSync(incompleteSkillFolder, { recursive: true });
-    writeFileSync(join(skillFolder, 'SKILL.md'), '---\nname: demos\ndescription: Demo runner\n---\n');
-
-    const adapter = createPiAdapter();
-    const compositeProfilePlan = adapter.createCompositeProfile(
-      { id: 'data_analyst', inherits: [], controls: {} },
-      { rootDirectory: join(root, 'composite'), profilePaths: [], profileFolders: [profileFolder] },
-    );
-    const launchPlan = adapter.createLaunchPlan(
-      compositeProfilePlan.compositeProfile,
-      { id: 'data_analyst', inherits: [], controls: { pi: { skills: ['user-skill'] } } },
-      [],
-      { profileFolders: [profileFolder] },
-    );
-
-    expect(launchPlan.args).toEqual([
-      '--prompt-template',
-      builtInOutfitterPromptTemplate,
-      '--skill',
-      builtInOutfitterSkill,
-      '--skill',
-      'user-skill',
-      '--skill',
-      skillFolder,
-    ]);
-  });
-
   it('reports invalid profile-bundled Pi resource paths', () => {
     const root = createTemporaryPiAdapterTestRoot('outfitter-pi-profile-resource-errors-');
     const adapter = createPiAdapter();
